+++ cmd/aitaskbuilder/batches.go
@@ -17,6 +17,7 @@ func NewBatchesCommand(client client.API, w io.Writer) *cobra.Command {
 
 	cmd.AddCommand(
 		NewBatchCreateCommand(client, w),
+		NewBatchInstructionsCommand(client, w),
 		NewGetBatchCommand(client, w),
 		NewGetBatchStatusCommand(client, w),
 		NewGetBatchesListCommand(client, w),